Add files via upload
This is a very common known game - "Stone Paper Scissor", designed with the help of HTML, CSS, and JS. (We can choose our option and from other side computer will generate options, further the result will come in the form of scores. Moreover we can reset the game score anytime.